Move classes of jflex-maven-plugin in jflex.maven.plugin.jflex
