+++ jflex-maven-plugin/src/main/java/jflex/maven/plugin/jflex/JFlexMojo.java
@@ -6,7 +6,7 @@
  * License: BSD                                                            *
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
-package de.jflex.plugin.maven;
+package jflex.maven.plugin.jflex;
 
 import static com.google.common.base.Strings.isNullOrEmpty;
