+++ jflex-maven-plugin/src/main/java/jflex/maven/plugin/jflex/LexSimpleAnalyzer.java
@@ -7,7 +7,7 @@
  * License: BSD                                                            *
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
-package de.jflex.plugin.maven;
+package jflex.maven.plugin.jflex;
 
 import com.google.common.io.Files;
 import java.io.File;
