+++ jflex-maven-plugin/src/test/java/jflex/maven/plugin/jflex/ClassInfoTest.java
@@ -6,7 +6,7 @@
  * License: BSD                                                            *
  *                                                                         *
  * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * */
-package de.jflex.plugin.maven;
+package jflex.maven.plugin.jflex;
 
 import static com.google.common.truth.Truth.assertThat;
